requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.6
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-tables2==2.3.1
django-wysiwyg==0.8.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
